Subject: packet serialization impl1 - flat buffer way

+++ GameServer/ServerPacketHandler.h
@@ -16,6 +16,6 @@ class ServerPacketHandler
 public:
 	static void HandlePacket(BYTE* buffer, int32 len);
 
-	static SendBufferRef Make_S_TEST(uint64 id, uint32 hp, uint32 power, vector<BuffData> buffs, wstring name);
+	static SendBufferRef Make_S_TEST(uint64 id, uint32 hp, uint16 power, vector<BuffData> buffs, wstring name);
 };
